+++ pkg/client/api/v0042/v0042.go
@@ -25,6 +25,7 @@ type ClientInterface interface {
 	JobInfoInterface
 	NodeInterface
 	PartitionInterface
+	StatsInterface
 }
 
 type SlurmClient struct {